Test that blogs are ordered by likes
This test is a bit long and crude in my opinion and highlights that
e.g. authentication should be moved into a cypress command, like the
material did. Still, it does validate what it should by creating two
blogs, liking one of them, refreshing and checking the order.

This marks my answer for exercise 5.22.